Move HRM Settings functionality to HRM module and optimize settings save functions to one
