+++ includes/framework/class-ajax.php
@@ -2,7 +2,6 @@
 
 namespace WeDevs\ERP\Framework;
 
-use WeDevs\ERP\Framework\Settings\ERP_Settings_General;
 use WeDevs\ERP\Framework\Traits\Ajax;
 use WeDevs\ERP\Framework\Traits\Hooker;
 
@@ -32,16 +31,50 @@ class Ajax_Handler {
      * @return void
      */
     public function init_actions () {
-        $this->action( 'wp_ajax_erp-settings-save', 'erp_settings_save_general' );
+        $this->action( 'wp_ajax_erp-settings-save', 'erp_settings_save' );
         $this->action( 'wp_ajax_erp-settings-get-data', 'erp_settings_get_general' );
-        $this->action( 'wp_ajax_erp-settings-workdays-get-data', 'erp_settings_get_workdays' );
-        $this->action( 'wp_ajax_erp-settings-workdays-save', 'erp_settings_save_workdays' );
-        $this->action( 'wp_ajax_erp-settings-leave-get-data', 'erp_settings_get_leaves' );
-        $this->action( 'wp_ajax_erp-settings-leave-save', 'erp_settings_save_leaves' );
-        $this->action( 'wp_ajax_erp-settings-miscellaneous-get-data', 'erp_settings_get_miscellaneous' );
-        $this->action( 'wp_ajax_erp-settings-miscellaneous-save', 'erp_settings_save_miscellaneous' );
-        $this->action( 'wp_ajax_erp-settings-get-hr-financial-years', 'erp_settings_get_hr_financial_years' );
-        $this->action( 'wp_ajax_erp-settings-financial-years-save', 'erp_settings_save_hr_financial_years' );
+        $this->action( 'wp_ajax_erp-settings-save-data', 'erp_settings_save' );
+    }
+
+    /**
+     * Save Settings Data For HRM's sections
+     *
+     * @since 1.8.6
+     *
+     * @return void
+     */
+    public function erp_settings_save() {
+        try {
+
+            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
+                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
+            }
+
+            $module  = sanitize_text_field( wp_unslash( $_POST['module'] ) );
+            $section = sanitize_text_field( wp_unslash( $_POST['section'] ) );
+
+            switch ( $module ) {
+                case 'general':
+                    $settings = ( new \WeDevs\ERP\Framework\Settings\ERP_Settings_General() );
+                    break;
+
+                case 'hrm':
+                    $settings = ( new \WeDevs\ERP\HRM\Settings() );
+                    break;
+
+                default:
+                    $settings = ( new \WeDevs\ERP\Framework\ERP_Settings_Page() );
+                    break;
+            }
+
+            $settings->save( $section );
+
+            $this->send_success([
+                'message' => __( 'Settings saved successfully !', 'erp' )
+            ]);
+        } catch ( \Exception $e ) {
+            $this->send_error( $e->getMessage() );
+        }
     }
 
     /**
@@ -58,7 +91,7 @@ class Ajax_Handler {
                 $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
             }
 
-            $settings = (new ERP_Settings_General());
+            $settings = ( new \WeDevs\ERP\Framework\Settings\ERP_Settings_General() );
             $settings->save();
 
             $this->send_success([
@@ -86,195 +119,6 @@ class Ajax_Handler {
             $data = erp_settings_get_general();
 
             $this->send_success( $data );
-        } catch (\Exception $e) {
-            $this->send_error( $e->getMessage() );
-        }
-    }
-
-    /**
-     * Get Settings Data For HR Workdays Section
-     *
-     * @since 1.8.6
-     *
-     * @return void
-     */
-    public function erp_settings_get_workdays() {
-        try {
-
-            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
-                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
-            }
-
-            $data = erp_settings_get_workdays();
-
-            $this->send_success( $data );
-        } catch (\Exception $e) {
-            $this->send_error( $e->getMessage() );
-        }
-    }
-
-    /**
-     * Save Settings Data For General Tab
-     *
-     * @since 1.8.6
-     *
-     * @return void
-     */
-    public function erp_settings_save_workdays() {
-        try {
-
-            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
-                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
-            }
-
-            erp_settings_save_workdays( $_POST );
-
-            $this->send_success([
-                'message' => __( 'Settings saved successfully !', 'erp' )
-            ]);
-        } catch ( \Exception $e ) {
-            $this->send_error( $e->getMessage() );
-        }
-    }
-
-    /**
-     * Get Settings Data For HR Workdays Section
-     *
-     * @since 1.8.6
-     *
-     * @return void
-     */
-    public function erp_settings_get_leaves() {
-        try {
-
-            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
-                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
-            }
-
-            $data = erp_settings_get_leaves();
-
-            $this->send_success( $data );
-        } catch (\Exception $e) {
-            $this->send_error( $e->getMessage() );
-        }
-    }
-
-    /**
-     * Save Settings Data For General Tab
-     *
-     * @since 1.8.6
-     *
-     * @return void
-     */
-    public function erp_settings_save_leaves() {
-        try {
-
-            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
-                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
-            }
-
-            erp_settings_save_leaves( $_POST );
-
-            $this->send_success([
-                'message' => __( 'Settings saved successfully !', 'erp' )
-            ]);
-        } catch ( \Exception $e ) {
-            $this->send_error( $e->getMessage() );
-        }
-    }
-
-    /**
-     * Get Settings Data For Miscellaneous
-     *
-     * @since 1.8.6
-     *
-     * @return void
-     */
-    public function erp_settings_get_miscellaneous() {
-        try {
-
-            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
-                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
-            }
-
-            $data = erp_settings_get_miscellaneous();
-
-            $this->send_success( $data );
-        } catch (\Exception $e) {
-            $this->send_error( $e->getMessage() );
-        }
-    }
-
-    /**
-     * Save Settings Data For Miscellaneous
-     *
-     * @since 1.8.6
-     *
-     * @return void
-     */
-    public function erp_settings_save_miscellaneous() {
-        try {
-
-            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
-                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
-            }
-
-            erp_settings_save_miscellaneous( $_POST );
-
-            $this->send_success([
-                'message' => __( 'Settings saved successfully !', 'erp' )
-            ]);
-        } catch ( \Exception $e ) {
-            $this->send_error( $e->getMessage() );
-        }
-    }
-
-    /**
-     * Get Settings Data For HR Financial years
-     *
-     * @since 1.8.6
-     *
-     * @return void
-     */
-    public function erp_settings_get_hr_financial_years() {
-        try {
-
-            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
-                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
-            }
-
-            $years = erp_get_hr_financial_years();
-
-            $this->send_success( $years );
-        } catch (\Exception $e) {
-            $this->send_error( $e->getMessage() );
-        }
-    }
-
-    /**
-     * Save Settings for HR Financial Years
-     *
-     * @since 1.8.6
-     *
-     * @return void
-     */
-    public function erp_settings_save_hr_financial_years() {
-        try {
-
-            if ( ! isset( $_POST['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['_wpnonce'] ), 'erp-settings-nonce' ) ) {
-                $this->send_error( __( 'Error: Nonce verification failed', 'erp' ) );
-            }
-
-            $inserted = erp_settings_save_leave_years( $_POST['fyears'] );
-
-            if ( is_wp_error( $inserted ) ) {
-                $this->send_error( __( $inserted->get_error_message(), 'erp' ) );
-            }
-
-            $this->send_success( [
-                'data'    => $inserted,
-                'message' => __( 'Settings saved successfully !', 'erp' )
-            ] );
         } catch (\Exception $e) {
             $this->send_error( $e->getMessage() );
         }
