+++ modules/hrm/includes/class-settings.php
@@ -141,9 +141,9 @@ class Settings extends ERP_Settings_Page {
         ];
         $fields = apply_filters( 'erp_settings_hr_section_fields', $fields, $section );
 
-        foreach ( $this->get_sections() as $section => $name ) {
-            if ( empty( $fields[ $section ] ) ) {
-                $fields = apply_filters( 'erp_settings_hr_section_fields', $fields, $section );
+        foreach ( $this->get_sections() as $sec => $name ) {
+            if ( empty( $fields[ $sec ] ) ) {
+                $fields = apply_filters( 'erp_settings_hr_section_fields', $fields, $sec );
             }
         }
